[PURPLE-12] 충돌 해결
