+++ src/main/java/com/pikachu/purple/domain/user/entity/User.java
@@ -1,6 +1,7 @@
 package com.pikachu.purple.domain.user.entity;
 
 import com.pikachu.purple.domain.user.enums.SocialLoginProvider;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,21 +12,38 @@ import lombok.NoArgsConstructor;
 public class User {
 
     private Long id;
-    private String nickname;
     private String email;
+    private String nickname;
+    private LocalDateTime registeredAt;
     private SocialLoginProvider socialLoginProvider;
 
     @Builder
     public User(
         Long id,
-        String nickname,
         String email,
+        String nickname,
+        LocalDateTime registeredAt,
         SocialLoginProvider socialLoginProvider
-    ) {
+    ){
         this.id = id;
-        this.nickname = nickname;
         this.email = email;
+        this.nickname = nickname;
+        this.registeredAt = registeredAt;
         this.socialLoginProvider = socialLoginProvider;
+    }
+
+    public static User create(Long id, String email, String nickname, SocialLoginProvider socialLoginProvider) {
+        return new User(
+            id,
+            email,
+            nickname,
+            LocalDateTime.now(),
+            socialLoginProvider
+        );
+    }
+
+    public void updateNickname(String nickname) {
+        this.nickname = nickname;
     }
 
 }
